refactor: extract lyrics visual editor config to separate module
