+++ src/components/generate-form/LyricsVisualEditor.tsx
@@ -45,13 +45,17 @@ import { VoiceInputButton } from "@/components/ui/VoiceInputButton";
 import { SectionTagSelector } from "./SectionTagSelector";
 import { TagBadge, TagList } from "@/components/lyrics/shared/TagBadge";
 import { cn } from "@/lib/utils";
-
-interface LyricSection {
-  id: string;
-  type: "intro" | "verse" | "chorus" | "bridge" | "outro" | "hook" | "pre" | "drop" | "breakdown";
-  content: string;
-  tags?: string[];
-}
+import { Plus, GripVertical, Trash2, Wand2, ChevronDown, ChevronUp, Tag, Music } from "@/lib/icons";
+import { Textarea } from "@/components/ui/textarea";
+import { toast } from "sonner";
+import {
+  SECTION_TYPES,
+  STRUCTURE_TEMPLATES,
+  parseLyrics,
+  sectionsToLyrics,
+  getCleanCharCount,
+  type LyricSection,
+} from "./lyricsVisualEditorConfig";
 
 interface LyricsVisualEditorProps {
   value: string;
@@ -59,209 +63,10 @@ interface LyricsVisualEditorProps {
   onAIGenerate?: () => void;
 }
 
-const SECTION_TYPES = [
-  {
-    value: "intro",
-    label: "Вступление",
-    icon: "🎬",
-    color: "bg-emerald-500/20 text-emerald-400 border-emerald-500/40",
-    gradient: "from-emerald-500/10 to-emerald-500/5",
-  },
-  {
-    value: "verse",
-    label: "Куплет",
-    icon: "📝",
-    color: "bg-sky-500/20 text-sky-400 border-sky-500/40",
-    gradient: "from-sky-500/10 to-sky-500/5",
-  },
-  {
-    value: "pre",
-    label: "Пре-припев",
-    icon: "⬆️",
-    color: "bg-amber-500/20 text-amber-400 border-amber-500/40",
-    gradient: "from-amber-500/10 to-amber-500/5",
-  },
-  {
-    value: "chorus",
-    label: "Припев",
-    icon: "🎵",
-    color: "bg-violet-500/20 text-violet-400 border-violet-500/40",
-    gradient: "from-violet-500/10 to-violet-500/5",
-  },
-  {
-    value: "hook",
-    label: "Хук",
-    icon: "🎤",
-    color: "bg-pink-500/20 text-pink-400 border-pink-500/40",
-    gradient: "from-pink-500/10 to-pink-500/5",
-  },
-  {
-    value: "bridge",
-    label: "Бридж",
-    icon: "🌉",
-    color: "bg-orange-500/20 text-orange-400 border-orange-500/40",
-    gradient: "from-orange-500/10 to-orange-500/5",
-  },
-  {
-    value: "drop",
-    label: "Дроп",
-    icon: "💥",
-    color: "bg-red-500/20 text-red-400 border-red-500/40",
-    gradient: "from-red-500/10 to-red-500/5",
-  },
-  {
-    value: "breakdown",
-    label: "Брейкдаун",
-    icon: "🔊",
-    color: "bg-cyan-500/20 text-cyan-400 border-cyan-500/40",
-    gradient: "from-cyan-500/10 to-cyan-500/5",
-  },
-  {
-    value: "outro",
-    label: "Концовка",
-    icon: "🔚",
-    color: "bg-slate-500/20 text-slate-400 border-slate-500/40",
-    gradient: "from-slate-500/10 to-slate-500/5",
-  },
-] as const;
-
-// Recommended structure templates
-const STRUCTURE_TEMPLATES = [
-  {
-    name: "Классическая",
-    icon: "🎼",
-    sections: ["intro", "verse", "chorus", "verse", "chorus", "bridge", "chorus", "outro"],
-  },
-  {
-    name: "Поп",
-    icon: "🎤",
-    sections: ["intro", "verse", "pre", "chorus", "verse", "pre", "chorus", "bridge", "chorus", "outro"],
-  },
-  {
-    name: "EDM",
-    icon: "🎧",
-    sections: ["intro", "verse", "build", "drop", "verse", "build", "drop", "breakdown", "drop", "outro"],
-  },
-  {
-    name: "Рэп",
-    icon: "🎙️",
-    sections: ["intro", "verse", "hook", "verse", "hook", "verse", "hook", "outro"],
-  },
-];
-
-// Helper function to parse lyrics text into sections
-function parseLyrics(text: string): LyricSection[] {
-  if (!text.trim()) return [];
-
-  const parsed: LyricSection[] = [];
-  const lines = text.split("\n");
-  let currentSection: LyricSection | null = null;
-
-  for (const line of lines) {
-    // Improved regex: matches [TAG] at start of line, potentially followed by content
-    // e.g., [Verse], [Verse 1], [Chorus, Powerful], or [Tag]Word
-    const sectionMatch = line.match(/^\[(\w+)(?:\s+\d+)?(?:,\s*[^\]]+)?\]/i);
-
-    if (sectionMatch) {
-      if (currentSection) {
-        parsed.push(currentSection);
-      }
-
-      const type = sectionMatch[1].toLowerCase() as LyricSection["type"];
-      const validType = SECTION_TYPES.find((t) => t.value === type) ? type : "verse";
-
-      // Get remaining content after the section tag (for cases like [Tag]Word)
-      const afterTag = line.slice(sectionMatch[0].length).trim();
-
-      // Extract any comma-separated tags from the section header like [Verse, Powerful]
-      const headerTags: string[] = [];
-      const fullMatch = line.match(/^\[([^\]]+)\]/);
-      if (fullMatch) {
-        const parts = fullMatch[1].split(",").map((p) => p.trim());
-        // First part is section type, rest are tags
-        headerTags.push(...parts.slice(1));
-      }
-
-      currentSection = {
-        id: `${validType}-${Date.now()}-${Math.random()}`,
-        type: validType,
-        content: afterTag, // Start with any content after the tag
-        tags: headerTags,
-      };
-    } else if (line.trim()) {
-      // Handle lines without section headers
-      if (!currentSection) {
-        // Create a default verse section if content appears before any section
-        currentSection = {
-          id: `verse-${Date.now()}-${Math.random()}`,
-          type: "verse",
-          content: "",
-          tags: [],
-        };
-      }
-
-      // Extract inline tags in brackets like [Powerful] or [Male Vocal]
-      const inlineTagMatches = Array.from(line.matchAll(/\[([^\]]+)\]/g));
-      const structureKeywords = ["verse", "chorus", "bridge", "intro", "outro", "hook", "pre", "drop", "breakdown"];
-
-      for (const tagMatch of inlineTagMatches) {
-        const tagContent = tagMatch[1];
-        const isStructure = structureKeywords.some((k) => tagContent.toLowerCase().startsWith(k));
-        if (!isStructure && currentSection.tags && !currentSection.tags.includes(tagContent)) {
-          currentSection.tags.push(tagContent);
-        }
-      }
-
-      // Extract parentheses tags like (ooh, aah), (harmony)
-      const parenTagMatches = Array.from(line.matchAll(/\(([^)]+)\)/g));
-      for (const tagMatch of parenTagMatches) {
-        const tagContent = tagMatch[1];
-        // Only add as tag if it looks like a production tag, not sung content
-        const looksLikeTag = /^[A-Za-z\s]+$/.test(tagContent) && tagContent.length < 30;
-        if (looksLikeTag && currentSection.tags && !currentSection.tags.includes(tagContent)) {
-          currentSection.tags.push(tagContent);
-        }
-      }
-
-      currentSection.content += (currentSection.content ? "\n" : "") + line;
-    }
-  }
-
-  if (currentSection) {
-    parsed.push(currentSection);
-  }
-
-  return parsed.length > 0 ? parsed : [];
-}
-
-// Helper function to convert sections back to lyrics text
-function sectionsToLyrics(sections: LyricSection[]): string {
-  return sections
-    .map((section) => {
-      // Use English tag names for Suno compatibility
-      const tagName = section.type.charAt(0).toUpperCase() + section.type.slice(1);
-      let content = section.content;
-
-      // Add tags inline if they exist and aren't already in content
-      if (section.tags && section.tags.length > 0) {
-        const existingTags: string[] = content.match(/\(([^)]+)\)/g) || [];
-        const tagsToAdd = section.tags.filter((tag: string) => !existingTags.includes(`(${tag})`));
-        if (tagsToAdd.length > 0) {
-          content = tagsToAdd.map((t) => `(${t})`).join(" ") + "\n" + content;
-        }
-      }
-
-      return `[${tagName}]\n${content}`;
-    })
-    .join("\n\n");
-}
-
-// Calculate character count excluding tags
-function getCleanCharCount(text: string): number {
-  return text
-    .replace(/\[.*?\]/g, "")
-    .replace(/\(.*?\)/g, "")
-    .trim().length;
+interface LyricsVisualEditorProps {
+  value: string;
+  onChange: (value: string) => void;
+  onAIGenerate?: () => void;
 }
 
 interface SortableLyricSectionProps {
